% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefactweak.R
\name{f901f1tnllk}
\alias{f901f1tnllk}
\title{nllk for 1-factor model with weak residual dep}
\usage{
f901f1tnllk(param, dstruct, iprfn = FALSE)
}
\arguments{
\item{param}{parameter vector (length 2\emph{d+2}(d-1))}

\item{dstruct}{= list with  data set $data on (0,1)^d
$edg1, $edg2 (d-1)x1 vectors for node labels of edges 1,...,d-1
for the tree for residual dependence,
$fam for copula family labels for the first tree}

\item{iprfn}{=indicator for printing of function and gradient (within NR iterations)}
}
\value{
nllk negative log-likelihood

grad gradient of nllk

hessian matrix of nllk
}
\description{
negative log-likelihood with gradient and Hessian computed in f90 for
copula from 1-factor/1-truncated vine (tree for residual dependence
conditional on a latent variable);
models included normal,t,BB1(r),BB8(r),gumbel,frank,survival gumbel
families for the frist tree and also the residual tree
}
