% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefactweak.R
\name{f901f1tproxynllk}
\alias{f901f1tproxynllk}
\title{approximate nllk for 1-factor copula model with weak res dep}
\usage{
f901f1tproxynllk(param, dstruct, iprfn = FALSE)
}
\arguments{
\item{param}{parameters in the 1-factor copula model with
weak res dep, vector of length 2*d-1}

\item{dstruct}{list with  data set $data on (0,1)^d}

\item{iprfn}{print gradient and hessian if true}
}
\value{
nllk: approximate nllk; grad: gradient of nllk;
hess: hessian matrix of nllk
}
\description{
This function computes the approxiamte nllk
for 1-factor copula model with the proxies plugged in
}
