% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefact.R
\name{f901fproxynllk}
\alias{f901fproxynllk}
\title{approximate nllk for 1-factor copula model}
\usage{
f901fproxynllk(param, dstruct, iprint = F)
}
\arguments{
\item{param}{parameters of dimension d
in the 1-factor copula model}

\item{dstruct}{list of input parameters: $lat: proxies of
the latent variables, $udata: observed data; $fam: copula
families for D linking copulas}

\item{iprint}{if T print gradient and Hessian}
}
\value{
nllk: approximate nllk; grad: gradient of nllk;
hess: hessian matrix of nllk
}
\description{
This function computes the approxiamte nllk
for 1-factor copula model with the proxies plugged in
}
