% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefact.R
\name{f901fproxynllk2}
\alias{f901fproxynllk2}
\title{approximate nllk for 1-factor copula model with copulas from
different families(2-parameter familes,like bb1, student-t)}
\usage{
f901fproxynllk2(param, dstruct, iprint = F)
}
\arguments{
\item{param}{2*d vector (par1,par2) for j-th linking copula}

\item{dstruct}{list of input parameters: $lat: proxies of
the latent variables, $udata: observed data; $fam: copula
families for D linking copulas}

\item{iprint}{if T print gradient and Hessian}
}
\value{
nllk: approximate nllk; grad: gradient of nllk;
hess: hessian matrix of nllk
}
\description{
This function computes the approxiamte nllk
for 1-factor copula model with the proxies plugged in
}
