% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifactor.R
\name{f90bifctFrkProxynllk}
\alias{f90bifctFrkProxynllk}
\title{approximate negtive log-likelihood for bi-factor}
\usage{
f90bifctFrkProxynllk(udata, th, vlat, grsize)
}
\arguments{
\item{udata}{input data in the u-scale. N*D}

\item{th}{parameters in the bi-factor model, vector of 2*D length(if all the
linking copulas are one-parameters copula)}

\item{vlat}{proxies of the latent}

\item{grsize}{group size in the model}
}
\value{
nllk negative log-likelihood
}
\description{
approximate negtive loglikielihood for bi-factor copula model with
all linking copulas Frank
}
