% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblique_factor.R
\name{f90str1proxynllk}
\alias{f90str1proxynllk}
\title{approximate nllk for oblique factor copula model}
\usage{
f90str1proxynllk(param, dstruct, iprfn = FALSE)
}
\arguments{
\item{param}{parameters in the oblique-factor copula model}

\item{dstruct}{list of input parameters:}

\item{iprfn}{print gradient and hessian if true}
}
\value{
nllk: approximate nllk; grad: gradient of nllk;
hess: hessian matrix of nllk
}
\description{
This function computes the approxiamte nllk
for oblique-factor copula model with the proxies plugged in
}
