% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefact.R
\name{find_start}
\alias{find_start}
\title{find a good starting point for 1-factor copula model}
\usage{
find_start(fam, tau, df)
}
\arguments{
\item{fam}{a vector of length d indicating the copula families}

\item{tau}{a vector of length d indicating the kendall's tau}

\item{df}{the fixed averaged degree of freedom for student-t copula}
}
\value{
st a vector of length 2*d for starting point
}
\description{
find a good starting point for optimization in
1-factor copula model with diffferent copula families
}
