% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifactor.R
\name{latUpdateBifctFrk}
\alias{latUpdateBifctFrk}
\title{compute new proxies for bi-factor copula models}
\usage{
latUpdateBifctFrk(udata, th, grsize, nq, xl, wl)
}
\arguments{
\item{udata}{input data in u-scale N*D}

\item{th}{parameters for linking copulas in the model}

\item{grsize}{group size in the bi-factor model}

\item{nq}{number of Gaussian points}

\item{xl}{Gaussian-Legendre points}

\item{wl}{Gaussian-Legendre weights}
}
\value{
v0: proxies of the global latent variable

vg: proxies of the local latent variables
}
\description{
latent update for the bi-factor copulas with all linking
copulas Frank
}
