% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefact.R
\name{latUpdateOnefct2}
\alias{latUpdateOnefct2}
\title{compute the new proxies corrected on the mean of observations}
\usage{
latUpdateOnefct2(th, udata, nq, xl, wl, family)
}
\arguments{
\item{th}{the estimated parameters in the 1-factor copula model}

\item{udata}{input N*D udata}

\item{nq}{number of  Gaussian-Legendre points}

\item{xl}{Gaussian quadrature points}

\item{wl}{Gaussian quadrature weights}

\item{family}{copula familes for d linking copulas}
}
\value{
lat: proxies of the latent variable
}
\description{
This function computes the proposed new proxies
in the 1-factor copula model with different copula families
}
