% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdhessian.R
\name{pdhessminb}
\alias{pdhessminb}
\title{optimize the negative log-likelihood}
\usage{
pdhessminb(
  param,
  objfn,
  ifixed,
  dstruct,
  LB,
  UB,
  mxiter = 30,
  eps = 1e-06,
  bdd = 5,
  iprint = F
)
}
\arguments{
\item{param}{starting point on entry}

\item{objfn}{function to be minimized with gradient and Hessian}

\item{ifixed}{= vector of length(param) of True/False, such that}

\item{dstruct}{= list with  data set and other variables used by objfn}

\item{LB}{= lower bound vector}

\item{UB}{= upper bound vector}

\item{mxiter}{= max number of iterations}

\item{eps}{= tolerance for Newton-Raphson iterations}

\item{bdd}{= bound on difference of 2 consecutive iterations}

\item{iprint}{= control on amount of printing}
}
\value{
fnval = function value at minimum
parmin = param for minimum
invh = inverse Hessian
iconv = 1 if converged, -1 for a boundary point, 0 otherwise
}
\description{
Version with ifixed as argument
}
