% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefact.R
\name{proxy1fct}
\alias{proxy1fct}
\title{parameter estimation using proxy method}
\usage{
proxy1fct(udata, family, start, LB, UB, xl, wl, iprint)
}
\arguments{
\item{udata}{N*D data in u-scale}

\item{family}{copula families of D linking copulas(the same), a number}

\item{start}{starting values for paramters}

\item{LB}{upper bound for parameters, length d}

\item{UB}{lower bound for parameters, length d}

\item{xl}{Gaussian-Legendre points}

\item{wl}{Gaussian-Legendre weights}

\item{iprint}{print the gradient and hessian if T}
}
\value{
mlpx1 estimated parameters from mean-proxy method
mlpx2 estimated parameters from new proxy method
proxyMean mean verison of proxies
proxyNew  new proposed proxies
}
\description{
parameter estimation using mean version of proxies in 1-factor
copula model
}
