% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefact.R
\name{proxy1fct2}
\alias{proxy1fct2}
\title{parameter estimation using proxy method}
\usage{
proxy1fct2(udata, fam, start, LB, UB, ifixed, xl, wl, iprint)
}
\arguments{
\item{udata}{N*D data in u-scale}

\item{fam}{copula families of D linking copulas, a vector of length d
now only allow 2: student-t, 4:gumbel  and 5: frank copulas}

\item{start}{starting values for paramters}

\item{LB}{upper bound for parameters, length d}

\item{UB}{lower bound for parameters, length d}

\item{ifixed}{if ifixed=T, the parameters at j-th position will be
fixed}

\item{xl}{Gaussian-Legendre points}

\item{wl}{Gaussian-Legendre weights}

\item{iprint}{print the gradient and hessian if T}
}
\value{
mlpx1 estimated parameters from mean-proxy method
mlpx2 estimated parameters from new proxy method
proxyMean mean verison of proxies
proxyNew  new proposed proxies
}
\description{
parameter estimation using mean version of proxies in 1-factor
copula model with different copula families
}
