% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefactweak.R
\name{proxy1fctweak}
\alias{proxy1fctweak}
\title{parameter estimation using proxy method}
\usage{
proxy1fctweak(udata, fam, start, LB, UB, xl, wl, iprint, ifixed)
}
\arguments{
\item{udata}{N*D data in u-scale}

\item{fam}{copula families of linking copulas}

\item{start}{starting values for paramters}

\item{LB}{upper bound for parameters, length d}

\item{UB}{lower bound for parameters, length d}

\item{xl}{Gaussian-Legendre points}

\item{wl}{Gaussian-Legendre weights}

\item{iprint}{print the gradient and hessian if T}

\item{ifixed}{paramters are set to be fixed}
}
\value{
mlpx1 estimated parameters from mean-proxy method
mlpx2 estimated parameters from new proxy method
proxyMean mean verison of proxies
proxyNew  new proposed proxies
}
\description{
parameter estimation using mean version of proxies in 1-factor
copula model
}
