% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifactor.R
\name{proxyBifact}
\alias{proxyBifact}
\title{parameter estimation in bi-factor copula model using proxy method}
\usage{
proxyBifact(udata, grsize, start, xl, wl, LB, UB, iprint)
}
\arguments{
\item{udata}{N*D data in u-scale}

\item{grsize}{group size in the model}

\item{start}{starting values for paramters}

\item{xl}{Gaussian-Legendre points}

\item{wl}{Gaussian-Legendre weights}

\item{LB}{upper bound for parameters, length d}

\item{UB}{lower bound for parameters, length d}

\item{iprint}{print the gradient and hessian if T}
}
\value{
mlpx1 estimated parameters from mean-proxy method
mlpx2 estimated parameters from new proxy method
proxyMean mean verison of proxies
proxyNew  new proposed proxies
condNum condition number for apa matrix
}
\description{
parameter estimation using mean version of proxies in
bi-factor copula model
}
