% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefact.R
\name{return_bound}
\alias{return_bound}
\title{return the optimization upper and lower bound for
the parameters in 1-factor copula model}
\usage{
return_bound(familyvec)
}
\arguments{
\item{familyvec}{a vector of length d for copula families}
}
\value{
UB1 upper bound for par1
UB2 upper bound for par2
LB1 lower bound for par1
LB2 lower bound for par2
}
\description{
upper and lower bound for parameters in 1-factor copula
model
}
