% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{sim1fact3}
\alias{sim1fact3}
\title{simulate data from 1-factor copula model with different
copula familes}
\usage{
sim1fact3(n, d, param, fam)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension}

\item{param}{parameter vector of dimension 2*d (par1,par2) for
d linking copulas, one-parameter set par2=0}

\item{fam}{family vector for d linking copulas
same index with the vinecopula package, 4:gumbel
5:frank copula;  2: student-t copula}
}
\value{
yy N*D simulate data in the u-scale

vv simulated latent variables
}
\description{
simulate data from 1-factor copula model
}
