% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simbifact2}
\alias{simbifact2}
\title{simulate data from bi-factor copula model}
\usage{
simbifact2(nn, grsize, cop = 5, param)
}
\arguments{
\item{nn}{sample size}

\item{grsize}{group size in the model}

\item{cop}{copula familes in the model}

\item{param}{paramters on the linking copulas}
}
\value{
zdata simulate data in the u-scale

v0 simulated glocal latent variables

vg simulate local latent variables
}
\description{
simulate data from bi-factor copula model
}
