% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simnestfact2}
\alias{simnestfact2}
\title{simulate data from nested oblique factor model}
\usage{
simnestfact2(nn, grsize, cop1, cop2, param)
}
\arguments{
\item{nn}{sample size}

\item{grsize}{group size in the model}

\item{cop1}{linking copula familes in T1}

\item{cop2}{linking copula families in each group}

\item{param}{paramters on the linking copulas}
}
\value{
zdata data in the u-scale

lat0 global latent variables

lat local latent variables
}
\description{
simulate data from nested oblique factor model
}
