% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simWeakRes}
\alias{simWeakRes}
\title{simulate data from 1-factor with weak res dependence}
\usage{
simWeakRes(
  n,
  d,
  thlow1 = 1.67,
  thupp1 = 5,
  thlow2 = 0.95,
  thupp2 = 2.95,
  fam1 = 4,
  fam2 = 5
)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension}

\item{thlow1}{lower bound for param in tree1}

\item{thupp1}{upper bound for param in tree1}

\item{thlow2}{lower bound for param in res tree}

\item{thupp2}{upper bound for param in res tree}

\item{fam1}{copula family in T1}

\item{fam2}{copula family in residual tree}
}
\value{
simdata simulate data in u-scale

thv paramters in the model
}
\description{
simulate data from 1-factor with weak residual
depdence, the weak dependence is modeled by 1-truncated
D-vine
}
