% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{uscore}
\alias{uscore}
\title{borrowed from packaage CopulaModel}
\usage{
uscore(data, aunif = -0.5)
}
\arguments{
\item{data}{input data}

\item{aunif}{set to be default value -0.5}
}
\value{
out data after transformation
}
\description{
transform data into u-scale
}
